Move CLI code review to extension code
Summary:
Doing a bit of clean-up in this diff by moving the code review bits I originally wrote from the webview to the extension code. This is mostly to help the mental model of where things should be located, and to facilitate similarly migrating the storage layer.

This isn't super clean because the storage still is managed via the `localStorageBackedAtom`, so multiple rounds of message passing are required to get everything synced up.

The eventual plan is to discard this in favor of running the review through DVSC / the ISL API. However, we don't have a consistent way to get a formatted review from the chat yet, so this is currently the best way to test UI bits consistently.

Reviewed By: evangrayk

Differential Revision: D82670744

fbshipit-source-id: 2d5fa8934cbe852c506399e22ca1ab37898ac400