cancell all GC operations before graceful restart
Summary:
## Background
Periodic GC runs on background server's threadpool. See:
https://www.internalfb.com/code/fbsource/[debecab6a176]/fbcode/eden/fs/service/EdenServer.cpp?lines=2988

By default, folly futures and thread pools do not provide a built-in way to forcibly stop or cancel a running task once it has started. We have to define a cancellation token that the background task checks periodically. When the token get canceled the task should stop

## This diff
Add `cancelAllGarbageCollections` to EdenServer and call it before graceful restrat

Differential Revision: D82591587

fbshipit-source-id: 316336486056ecb436f874556bd349d2e3763802