implement a source cancellation for background GC
Summary:
## Background
Periodic GC runs on background server's threadpool. See:
https://www.internalfb.com/code/fbsource/[debecab6a176]/fbcode/eden/fs/service/EdenServer.cpp?lines=2988

By default, folly futures and thread pools do not provide a built-in way to forcibly stop or cancel a running task once it has started. We have to define a cancellation token that the background task checks periodically. When the token get canceled the task should stop

## This diff
Implement the `folly::CancellationSource` in EdenServer

Reviewed By: genevievehelsel

Differential Revision: D82588941

fbshipit-source-id: d0cf2efaccec5a5ace4cd53c1f99aabe2febb290