sql_ext: Move error logging to `inspect_err` block from Retry
Summary:
## This stack
Minor improvements to the SQL telemetry.

In these next few diffs, I'll improve the way we log errors and query completion time, so we can

- Better understand retries and how often they bubble up to the caller
- compare the performance of Consistent read queries with reads from the primary.

## This diff
Move the error logging to the `inspect_err` block for the queries that have retry enabled.
This will let us log the query attempt number with the error, to get a better sense of how often the errors are masked or thrown to the caller.

Reviewed By: lmvasquezg

Differential Revision: D81749532

fbshipit-source-id: a7ce3e1352144c838b1c1c6e4ea61b858f535149