sql_ext: Introduce ConsistentRead granularity for to track latency and errors
Summary:
## This stack
Minor improvements to the SQL telemetry.

In these next few diffs, I'll improve the way we log errors and query completion time, so we can

- Better understand retries and how often they bubble up to the caller
- compare the performance of Consistent read queries with reads from the primary.

## This diff
Introduce a new granularity, `ConsistentRead`, to track telemetry about **an entire consistent read query**, which might involve many individual read queries to replicas and a single fallback to the primary.

`ConsistentReadQuery` will be used for the individual queries that are part of a consistent read.

## Why?
I want to be able to compare the performance of `query_with_consistency` with the current behaviour of falling back to master directly.

Reviewed By: andreacampi

Differential Revision: D82813554

fbshipit-source-id: b76e8b15f8e293041fd8aa1e1cc197e085c53c5f