scs: add hg-mutation format for hg mutation history method
Summary:
The current hg mutation history method only returns the set of historical commits (not in an particular order).  If we want to recover the full mutation history graph, that's not possible.

Add a new mutation history format (`HgMutation`) that returns a list of all of the mutations, including predecessors and successor, as well as splits, and the metadata about who modified the commit and and when.

Reviewed By: gustavoavena

Differential Revision: D83000813

fbshipit-source-id: 4f778b47623a10f4cea8368e6c6e6fd716b2691a