python-modules: use cpython rust_library to match desired Python version
Summary:
Previously, we use a `python_binary` buck target to provide the desired Python
version by providing a Python interpreter. This is a unique customization in
the whole Sapling code base. This diff changes it to use the `cpython`
`rust_library` instead to decide the Python version. The `cpython`
`rust_library` is used in many other places so hopefully this is more
consistent (and more reliable since we no longer need a Python executable, but
only the libpython).

Reviewed By: muirdm

Differential Revision: D83021220

fbshipit-source-id: 660eea744334f95734ebc94b6bdfa4ceac53ee1e