push: don't update visibility when pulling bookmark
Summary: We do a `repo.pull()` during push to make sure we have pulled the remote bookmark before we try pushing to it since this affects draft/public. However, this pull also had a side effect of adding the about-to-be-moved bookmark's node to visible heads, causing it to stay visible even when it shouldn't.

Reviewed By: zzl0

Differential Revision: D83163250

fbshipit-source-id: 267049e83fc2facfea24bce6e5168e60f5effbf9