sparse: introduce from_profiles
Summary:
# This diff

introduces a new method, from_profiles, that can be used to construct a root that contains only a series of `%include <profile>` entries in it.

# Context

This is needed because we're going to start supporting the ability to enable multiple filters with FilteredFS, and therefore the FilteredFS FFI crate needs to be able to create a sparse matcher that contains entries from multiple paths in the repo.

Reviewed By: muirdm

Differential Revision: D83016996

fbshipit-source-id: 26f8e184afa43dd7921f94b5044d6305be119212