edenfs-client: refactor get_active_filter_id to use Filters
Summary:
# This diff

Refactors the existing get_active_filter_id method in the following ways:

- Makes the method return a FilterId instead of a String
- Moves the logic for reading the filter file into a separate method
- Conditonally construct a V1 or Legacy Filter based on which default_filter_version was used when constructing the FilterGenerator
- Return errors for invalid filter files instead of simply warning and returning None

# Context

Some workflows now need to apply multiple FilteredFS filters at once. This could theoretically be achieved by slightly changing the FilterID format to support multiple comma-separated paths as the first component (i.e. `<path1,path2,path3>:<commit_id>`). However, this would mean that the FilterID could grow arbitrarily long in length (and consist of many hundreds/thousands of bytes of data). This is problematic since FilterIDs are inserted into EdenFS ObjectIDs, and ObjectID [de]serialization could become very expensive if FilterIDs grow too long.

Therefore, we're introducing a brand new FilterID format and storage mechanism in Sapling. Sapling will still track path/commit information inside filters, but it will hash all this data and use the hash of the data as the FilterID. We will also add a version to the FilterID so we know the [de]serialization procedure for parsing/storing the FilterID in case FilterID formats change again in the future. [De]serialization for V1 filters will be handled by serde/mincode.

The actual content of the filter (which was hashed to produce the FilterID) will be stored on disk in the .hg directory in an indexedlog. The indexedlog will be indexed by the partial blake3 hash of the Filter content. The indexedlog will be a permanent indexedlog (i.e. it won't rotate), so Filter information will be permanently stored during the lifetime of the repository.

Reviewed By: muirdm

Differential Revision: D81462082

fbshipit-source-id: c9e52bb102acbccffc5d8431588c1a5366bfcf1c