edenfs-client: killswitch filter storage separately from filter version
Summary:
# This diff

Refactors the filter indexedlog killswitch logic so that the killswitch is controlled with a separate config.

# Context

Using the same config for the FilterVersion and Indexedlog killswitch is a bad idea. If we go from V1 filters to Legacy filters, the indexedlog killswitch would activate. This means that any users that had V1 filters stored on disk would no longer be able to access Filter content for those filters, rendering their repos useless.

Instead, we want the killswitch to be separate, so that we can turn off indexedlog only if construction of the indexedlog is causing issues. In most cases, we would want to roll back to Legacy filters but keep the indexedlog available so that any previously serialized V1 filters could still be interpreted by Sapling/EdenFS.

Reviewed By: muirdm

Differential Revision: D83016998

fbshipit-source-id: 252652e32c14be0dc003fae96c9be3d7cc654b97