edenfs-client: add default FilterVersion to FilterGenerator
Summary:
# This diff

Up until the next diff, everything introduced in this stack was more or less shielded from production since it was not used in any business logic. The next diff alters a core method that's already used for FilteredFS, so I wanted to gate those changes from production as best as possible.

This diff adds a default_filter_version field to the FilterGenerator that is used to determine which FilterVersion get_active_filter_id should construct when parsing the .hg/sparse file and generating an active FilterId.

By default, we'll use the Legacy filter type so that we can slowly roll out V1 Filters to production.

# Context

Some workflows now need to apply multiple FilteredFS filters at once. This could theoretically be achieved by slightly changing the FilterID format to support multiple comma-separated paths as the first component (i.e. `<path1,path2,path3>:<commit_id>`). However, this would mean that the FilterID could grow arbitrarily long in length (and consist of many hundreds/thousands of bytes of data). This is problematic since FilterIDs are inserted into EdenFS ObjectIDs, and ObjectID [de]serialization could become very expensive if FilterIDs grow too long.

Therefore, we're introducing a brand new FilterID format and storage mechanism in Sapling. Sapling will still track path/commit information inside filters, but it will hash all this data and use the hash of the data as the FilterID. We will also add a version to the FilterID so we know the [de]serialization procedure for parsing/storing the FilterID in case FilterID formats change again in the future. [De]serialization for V1 filters will be handled by serde/mincode.

The actual content of the filter (which was hashed to produce the FilterID) will be stored on disk in the .hg directory in an indexedlog. The indexedlog will be indexed by the partial blake3 hash of the Filter content. The indexedlog will be a permanent indexedlog (i.e. it won't rotate), so Filter information will be permanently stored during the lifetime of the repository.

Reviewed By: muirdm

Differential Revision: D81724681

fbshipit-source-id: 8b8daf01190140d6601e89059ca3fbafc6f0cc38