edenfs-client: add Filter struct and constructors
Summary:
# This diff

Introduces the Filter struct which will be used to represent a Filter that's deserialized (or newly created and needs to be serialized). V1 Filter construction will eventually enforce that the Filter content is stored in the filter storage indexedlog, but that's added as a TODO for now since the diff would be very large if this was all in a single diff.

Note: I should likely allow for Filter deserialization to construct a FilterRef which just stores references to the underlying Filter content in the Indexedlog and can convert the Filter bytes to concrete types (or slices of concrete types) on the fly. However, for simplicity, I made the Filter implementation copy the data into typed fields for now.

# Context

Some workflows now need to apply multiple FilteredFS filters at once. This could theoretically be achieved by slightly changing the FilterID format to support multiple comma-separated paths as the first component (i.e. `<path1,path2,path3>:<commit_id>`). However, this would mean that the FilterID could grow arbitrarily long in length (and consist of many hundreds/thousands of bytes of data). This is problematic since FilterIDs are inserted into EdenFS ObjectIDs, and ObjectID [de]serialization could become very expensive if FilterIDs grow too long.

Therefore, we're introducing a brand new FilterID format and storage mechanism in Sapling. Sapling will still track path/commit information inside filters, but it will hash all this data and use the hash of the data as the FilterID. We will also add a version to the FilterID so we know the [de]serialization procedure for parsing/storing the FilterID in case FilterID formats change again in the future. [De]serialization for V1 filters will be handled by serde/mincode.

The actual content of the filter (which was hashed to produce the FilterID) will be stored on disk in the .hg directory in an indexedlog. The indexedlog will be indexed by the partial blake3 hash of the Filter content. The indexedlog will be a permanent indexedlog (i.e. it won't rotate), so Filter information will be permanently stored during the lifetime of the repository.

Reviewed By: muirdm

Differential Revision: D81551189

fbshipit-source-id: eab3a63a957c54e7c7e5b54ed89373e7862d5b69