FilterUtils: fix bug when FilterId contains null byte
Summary:
# This diff

Ensures that the entire FilterId is parsed regardless of whether it contains a null byte

# Context

Previously, we were relying on FilterIds being valid UTF8 and not containing nullbytes. Now that this changed, we need to support filters containing random null bytes in them.

This method was previously using a string constructor that parsed bytes until it found a null byte. Now we'll ensure that we parse the FilteredRootId until we reach the very end.

Reviewed By: muirdm

Differential Revision: D83114251

fbshipit-source-id: c4c50d49966a2debaf4aef5a094820157c7ce891