edenfs-client: introduce FilterVersion
Summary:
# This diff

Introduces FilterVersion, a struct for describing which FilterID/Filter version is being used to store a filter. There are currently 2 FilterVersions:

- Legacy: This is the old FilterID format that encodes filter information directly in the ID. It consists of `<filter_path>:<commit_id>`
- V1: A new Filter/FilterID format that stores the content of the Filter on disk in an indexedlog and uses the partial blake3 hash of Filter content as an index.

# Context

Some workflows now need to apply multiple FilteredFS filters at once. This could theoretically be achieved by slightly changing the FilterID format to support multiple comma-separated paths as the first component (i.e. `<path1,path2,path3>:<commit_id>`). However, this would mean that the FilterID could grow arbitrarily long in length (and consist of many hundreds/thousands of bytes of data). This is problematic since FilterIDs are inserted into EdenFS ObjectIDs, and ObjectID [de]serialization could become very expensive if FilterIDs grow too long.

Therefore, we're introducing a brand new FilterID format and storage mechanism in Sapling. Sapling will still track path/commit information inside filters, but it will hash all this data and use the hash of the data as the FilterID. We will also add a version to the FilterID so we know the [de]serialization procedure for parsing/storing the FilterID in case FilterID formats change again in the future. [De]serialization for V1 filters will be handled by serde/mincode.

The actual content of the filter (which was hashed to produce the FilterID) will be stored on disk in the .hg directory in an indexedlog. The indexedlog will be indexed by the partial blake3 hash of the Filter content. The indexedlog will be a permanent indexedlog (i.e. it won't rotate), so Filter information will be permanently stored during the lifetime of the repository.

Reviewed By: muirdm

Differential Revision: D81462083

fbshipit-source-id: 5a290341b45770c13fa84d1108e8ca596946ed15