edenfs-client: add Filter serialization
Summary:
# This diff

Derives [de]serialization for storing Filter objects. Note that Filters are stored in the Filter store indexedlog as:

[FilterId Index][mincode serialized Filter struct]

Where there must be at least 1 path stored along with the commit id.

# Context

Some workflows now need to apply multiple FilteredFS filters at once. This could theoretically be achieved by slightly changing the FilterID format to support multiple comma-separated paths as the first component (i.e. `<path1,path2,path3>:<commit_id>`). However, this would mean that the FilterID could grow arbitrarily long in length (and consist of many hundreds/thousands of bytes of data). This is problematic since FilterIDs are inserted into EdenFS ObjectIDs, and ObjectID [de]serialization could become very expensive if FilterIDs grow too long.

Therefore, we're introducing a brand new FilterID format and storage mechanism in Sapling. Sapling will still track path/commit information inside filters, but it will hash all this data and use the hash of the data as the FilterID. We will also add a version to the FilterID so we know the [de]serialization procedure for parsing/storing the FilterID in case FilterID formats change again in the future. [De]serialization for V1 filters will be handled by serde/mincode.

The actual content of the filter (which was hashed to produce the FilterID) will be stored on disk in the .hg directory in an indexedlog. The indexedlog will be indexed by the partial blake3 hash of the Filter content. The indexedlog will be a permanent indexedlog (i.e. it won't rotate), so Filter information will be permanently stored during the lifetime of the repository.

Reviewed By: muirdm

Differential Revision: D81551192

fbshipit-source-id: 86b39d222499adc875672c203b2a750bb6d82bf1