Pull recommended bookmarks
Summary:
There is a vast collection of remote bookmarks, and users are only by default subscribed to a small subset (```sl bookmarks --list-subscriptions```). A subset of this subscriptions appears in the bookmarks list.

For recommended bookmarks, we will check the subscriptions list to see if the recommended bookmark is already subscribed to. If not, we pull it with ```sl pull -B```, which adds it to the subscribed list. As per [the sapling Wiki](https://www.internalfb.com/wiki/Sapling/Advanced/Documentation/Overview/push-pull/) and my testing, this allows the bookmark to be updated with future ```sl pull```. Another option considered would be to add it to the remotenames.selectivepulldefaults and run a pull. This requires a general pull (which also adds the bookmarks to the subscription list but is more than needed), and requires modifying the config.

Conveniently, the subscribed bookmarks list acts as a cache, so we do not need to pull recommended bookmarks more than once per user. Additionally, this only happens once on startup.

Reviewed By: evangrayk

Differential Revision: D83014537

fbshipit-source-id: a83e5dde5e0f12bae66fdbf2d38f4b9c43d53654