code_tenting: Add RestrictedPaths to DerivationContext
Summary:
## This stack
Working on a prototype for server access logging for the code tenting workstream in the focus room.

## This diff
Add `RestrictedPaths` to `DerivationContext`.

## Why?
For every manifest that we'll support, we'll need to read the config during derivation to check if any of the manifests belong to a restricted path and, if it does, store this id in the `RestrictedPathsManifestIdStore`.

You can see how it will be used in D83012636.

Reviewed By: markbt

Differential Revision: D83157454

fbshipit-source-id: 24523bc0fe309f7b567a6ce2ed318507099f9dc1