code_tenting: Add RestrictedPaths repo facet
Summary:
## This stack
Working on a prototype for server access logging for the code tenting workstream in the focus room.

## This diff
Introduce the `RestrictedPaths` facet, which will store the `RestrictedPathsConfig`, containing all the restricted paths and their corresponding ACLs, and a `RestrictedPathsManifestIdStore`, providing an interface to both read and write the manifest ids from these paths.

It will be used in mercurial manifest derivation in D83012636.

Reviewed By: markbt

Differential Revision: D83010639

fbshipit-source-id: 0306a2d320a697ac55e1a339abe1c58008fb2648