Add basic DebugCommentGroup
Summary:
Now that we have the notion of separate InlineCommentGroups, we can add additional groups as useful, which is powerful.

Here we define a new group that loads JSON from the currently active file editor, and interprets that as comment data and makes the relevant comments in the group.

This way we can just edit some JSON to play around with comments, and test how realigning and other edge cases work, without needing to go to phabricator and make a real comment.

I only enable this feature in dev builds.

Reviewed By: nsblake

Differential Revision: D82684690

fbshipit-source-id: f21f5c562f80e0e7ed188c02634664c33fb4899c