Add tracking to inline comment groups
Summary:
Pass a tracker to the comment manager, and allow comment groups to access it to expose a `track()` method. Then, the inset webview message handler can get a reference to the group to do track calls. We also typically have a comment as a reference, so we can pass the comment ID if it exists.

The benefit of this approach is that inline comment groups can redefine their own track call to pass additional extras. For example, the DiffCommentGroup can pass the author ID or diff ID or whatever to all track calls within the group.

Reviewed By: nsblake

Differential Revision: D83073635

fbshipit-source-id: f42c68e9188cedf2a40cebf8cb50b0677031e3f5