Add messages for loading / error states
Summary:
Besides the small symbolic indicator, we should provide more robust messaging for the other states of the code review (error / loading).

If an error occurs, we can use the ErrorNotice component to render a collapsible stack trace. Doing this also helped me realize that I had a bug in my state logic.

If the review is in progress, I currently just show a generic loading message, but maybe we can tune this once we have proper DVSC reviews hooked up.

Likewise, we should show some other message when the review completes successfully but no comments are returned. The wording here can definitely be tweaked.

Reviewed By: evangrayk

Differential Revision: D83187791

fbshipit-source-id: 3f0098f702948c0881d39234a1e60863678c7355