Hook up runAICodeReview to AIReviewCommentGroup
Summary:
With the new comment architecture in place, we want `runCodeReview` (renamed to `runAICodeReview`) to use `AIReviewCommentGroup` to populate comment insets on top of files, instead of using the old `FirstPassCodeReviewInlineController`. This diff adds all of the necessary plumbing to make that work.

Later on, we'll replace `runAICodeReview` with a call to DVSC to perform the actual review.

Reviewed By: evangrayk

Differential Revision: D83074844

fbshipit-source-id: d746545ed12fe3d9f442030dad2ff7ff5c2446ed