Upsell for AI code review
Summary:
The current ISL UX for AI code review is doing double duty as both as an upsell and a status indicator. Let's separate these out by adding a dedicated upsell component. We can move this component to the top of the form so that it's more visible.

Afterwards I'll convert the existing status component into something where we can show a summary of review comments, on the path to a more permanent "signal box"-like UX.

(Of course, users will need to be in the [proper GK](https://www.internalfb.com/intern/gatekeeper/projects/isl_ai_first_pass_code_review/) to see this upsell in the first place).

Reviewed By: evangrayk

Differential Revision: D83086533

fbshipit-source-id: 85de62a43892de4276755ff0e3811974f50ca1ff