aosp: implement streaming changelog on slapi
Summary:
In our last hackathon we worked on deprecating wireproto, but one of the big blockers was AOSP clones regression when we used commit graph segments. Therefore, we decided to migrate the streaming changelog endpoint over to SLAPI.

This is a first draft that mimics  what the wireproto endpoint was doing, still need to do the client side but figured I'd put this up for review for now since I'm working on this on my bits of free time.

Reviewed By: muirdm

Differential Revision: D77028873

fbshipit-source-id: afab86424ecd192071d42bfd62a1a8a405f664ac