debugconflictcontext: add command to give context for file conflicts
Summary:
This command's purpose is to give useful context for each file with merge conflicts. Currently it tries to guess the commit that introduced the conflicts on the "local" side (i.e. the rebase destination side), and if it finds something then reports the commit message and diff of that commit. It also reports the diff and commit message of the "other" side (i.e. the rebase source).

I'm sure it has some details wrong and doesn't work in call cases, but it is a starting point we can iterate on.

Reviewed By: zzl0

Differential Revision: D83270106

fbshipit-source-id: b87771f5ff7f768fca7a5aa09196d683f2df232c