Add smart action for reviewing code
Summary:
For the purposes of discovery, we want to add first-pass code review as a Smart Action in the ISL. This is relatively easy to do, following the currently-established pattern for the Smart Actions menu.

However, we definitely want to re-think the design/UX of smart actions to better upsell code review (and other smart actions). This isn't a very clean or scalable way to add new smart actions, and it actually lacks in potential discoverability due to requiring a click to show the available actions.

evangrayk and I have some ideas to improve this UX, but for now we can just add the review code action so that we have an additional entrypoint.

Reviewed By: evangrayk

Differential Revision: D83280789

fbshipit-source-id: 09c8ada7f1eb2631e87abd041f669420d61b3aae