sparse: use Rust sparse matcher directly
Summary:
Hook things up so we pass the Rust sparse::Matcher directly to Python. Previously we decomposed the Rust matcher into underlying treematchers, and then Python reassmbled things. This was so Python could maintain the "explain" behavior, which mostly lived in Python. I am consolidating on the Rust explain logic, so we can dispense with the treematcher stuff.

The "real" reason I want this is so I can add more complex logic to the Rust sparse::Matcher, which would have otherwise been very difficult due to the assumption that sparse::Matcher is made up of a simple list of treematchers.

Reviewed By: quark-zju

Differential Revision: D83518273

fbshipit-source-id: bd1581e6c6f2d3518466752717f5edbc3f29e974