sparse: add new "required" flag for v2 profiles
Summary:
If set, "required" causes top-level v2 sparse profiles to be AND'd with the other top-level profiles instead of OR'd. This makes it possible to have a v2 profile whose only purpose is to exclude things (i.e. a global exclude list that applies to all other enabled filters).

This change would have been simple except for the "explain" logic which clarifies why the sparse profile matches or doesn't match. I incorporated the new boolean logic into the explanation using "AND(...)" syntax so you can see how the results from each top level profile are combined.

Reviewed By: MichaelCuevas

Differential Revision: D83518270

fbshipit-source-id: a7ee396a3fe664176c369d19bb89bcc29f3e99db