sparse: improve sparse matcher Rust explanation
Summary:
Juice up the Rust matcher's explanation logic to cover what the Python unionmatcher includes. Basically, give info for each sub-matcher that matched. I wrapped it all in an "OR()" to make the logic clear. This will come in handy later when I introduce "AND()".

I'm going to get rid of the Python explain logic so we can return an opaque Rust matcher to Python for more flexibility. This is step one.

Reviewed By: quark-zju

Differential Revision: D83518272

fbshipit-source-id: e704f6f8d2ddc6fdd302d98d2b3c4456f5b8e8cf