integration: enables V1 filters when running FilteredHg integration tests
Summary:
# This diff

Sets the appropriate Sapling configs for enabling V1 filters when running integration tests

# Context

We'd like to start using V1 filters in production, so we need to make sure they're tested sufficiently. Note: this uncovered a bug when enabling multiple filters at once. The bug is fixed in subsequent diffs in this stack.

Reviewed By: muirdm

Differential Revision: D83116060

fbshipit-source-id: 477b3a696a46b5846af4fb60707f826f4a4eeabd