add a new optional field to silent daemon exit logs
Summary:
## Context:
The silent exits are reported in edenfs_events table in scuba. However, we need to add some more data to these events therefore we can determine whether the silent exit was due to memory pressure or some other issues such as reboot. We have `log show` shell command on mac that log all the memory pressure events. If we search this log show between the last time that the old eden was alive and the start up time of the new eden, maybe we can find the memory pressure and log it into scuba

## This diff:
Add a new column for memory pressure in edenfs_events table

Reviewed By: muirdm

Differential Revision: D81367865

fbshipit-source-id: 50e279d51166dddddd9c052ee2560ffe9c453ba1