add a config for adding log show to silent exit logs
Summary:
## Context:
The silent exits are reported in edenfs_events table in scuba. However, we need to add some more data to these events therefore we can determine whether the silent exit was due to memory pressure or some other issues such as reboot. We have `log show` shell command on mac that log all the memory pressure events. If we search this log show between the last time that the old eden was alive and the start up time of the new eden, maybe we can find the memory pressure and log it into scuba

## This diff:
Add a new config to control roll out for this new column. `log show` could be expensive we should control the roll out of this new column, and close it if it break something

Differential Revision: D81349501

fbshipit-source-id: be3d843b3a852df5e07076f00da15c97d4828029