Add isMemoryPressureInLogShow to silent exit logs
Summary:
## This diff:
Run `log show` command and search the result to confirm if eden was killed because of a memory pressure or not. The `log show` is expensive the next diff move this part into background and it won't block eden start up.

## Context:
The silent exits are reported in edenfs_events table in scuba. However, we need to add some more data to these events therefore we can determine whether the silent exit was due to memory pressure or some other issues such as reboot. We have `log show` shell command on mac that log all the memory pressure events. If we search this log show between the last time that the old eden was alive and the start up time of the new eden, maybe we can find the memory pressure and log it into scuba

Reviewed By: muirdm

Differential Revision: D81362268

fbshipit-source-id: fa25836e7a048ff45bf1262d1754a40f04078331