move the start up time consuming logging in background
Summary:
## Context:
The silent exits are reported in edenfs_events table in scuba. However, we need to add some more data to these events therefore we can determine whether the silent exit was due to memory pressure or some other issues such as reboot. We have `log show` shell command on mac that log all the memory pressure events. If we search this log show between the last time that the old eden was alive and the start up time of the new eden, maybe we can find the memory pressure and log it into scuba

## This diff:
`isMemoryPressureInLogShow()` is expensive because it call a shell command `log show`. This diff moves this part into background and it won't block eden start up.

Reviewed By: muirdm

Differential Revision: D82869406

fbshipit-source-id: 2090d98f5f994f7b501126ba73cb2e9b09c78d41