Remove "devmate" references in smart actions component
Summary:
We reference Devmate a LOT in the smart actions menu component, but this component lives in OSS. As a short-term fix, let's use more generic terminology for the AI-powered actions. Long-term, we should move the source-of-truth for Meta-internal actions to internal code, so that we can continue to use more specific wording.

There are still some other places in OSS code where "Devmate" is referenced, but this is a start.

Reviewed By: stbfb

Differential Revision: D83501141

fbshipit-source-id: 12ee2d5c6c26ea466684e58234868d9f9b468446