check nullptr when stopping FAM
Summary:
## Context
A FAM user found that `eden fam stop` could crash edenfs. This is because when there is not fam started successfully, `famProcess_` is not initialized correctly. Trying to move something out from it will cause a crash.

In the user's case, FAM was actually not started successfully. So `famProcess_` was a nullptr when user tried to stop FAM.

FAM was not started successfully because the user was using the wrong flag from the outdated wiki. The user didn't realized that because edenfs rust cli command would just exit silently without printing an error message. This is fixed in the previous diff.

## This Diff
We check if `famProcess_` is initialized first. If it's not, we just throw an error since there is nothing we can clean from the privHelper's perspective.

Reviewed By: ViniGupta08

Differential Revision: D83604361

fbshipit-source-id: 6af1678e62dfda971ca9c8d5b8cbd8c438e67236