add a new argument to eden debug inode
Summary:
We have `DIS_COMPUTE_ACCURATE_MODE` flag for this command. See here:
https://www.internalfb.com/code/fbsource/[309926dc3959]/fbcode/eden/fs/service/EdenServiceHandler.cpp?lines=5091-5095
But this flag was not availble in `eden debug inode`
Add the flag to this debug command with `--show-permisison`

This flag can show the inode permission. Without this flag the `entryInfo.mode()` is only dtype and doesn't have the permission part of the file mode.

Differential Revision: D83622075

fbshipit-source-id: ee001ddd282d35cc1a3ffc6bc6551ad2c3635ca1