Add FakeBackingStore co_getBlob
Summary:
In this diff, add a coroutine implementation of `getBlob` to `FakeBackingStore` -- `co_getBlob`
This is working towards every child of `BackingStore` to have a `co_getBlob` implementation so we can make it a required function in the base class.

Notes:
I'm planning to pause the coroutine migration at `StoredObject` boundary. It uses futures to mark if a blob is ready or not. I'll revisit this in the future.

Differential Revision: D82578003

fbshipit-source-id: 939752f29054e911518e66fe8ae0a700f5094ed2