add cli argument to not require the x-client-info header
Summary:
With hg we control whether users pass the header or not tightly. We can't do that with git to a reasonable degree right now. That's why the mononoke-git server already doesn't require it.

Let's make this a cli argument as we plan to reuse lfs_server's code as much as possible for mononoke-git-lfs-server deployment - we can't just hardcoded it in code if we want this.

Reviewed By: YousefSalama

Differential Revision: D83755072

fbshipit-source-id: bc3d5c881881b18907d6e2e7a4fc2ef621252e27