Only show code review UX on VS Code
Summary:
Code review is (currently?) only available through DVSC, so we should only show the upsell and status components if we're on VS Code.

Originally I was going to move all of the first pass components to vscode/webview (which we should probably still do), but it's a little more complex since we're using Internal imports from isl/src. A quicker way is to just check the platform name in the components themselves.

Reviewed By: evangrayk

Differential Revision: D83767266

fbshipit-source-id: 0388ba4f2eef19dacc6b3c8aa0f080a2bcde7adc