Reduce the number of syscalls we make in the traversal benchmark
Summary:
### Summary

This diff aims to optimize the performance of the traversal benchmark in the `eden debug bench` command by reducing the number of syscalls made during the traversal process.

The main change is in the `traversal.rs` file, where the `entry.file_type()` method is called only once to get the file type, and the `is_dir()` and `is_symlink()` methods are called on the resulting `FileType` object instead of the `Path` object. This reduces the number of syscalls required to determine the file type.

Reviewed By: akushner

Differential Revision: D83882516

fbshipit-source-id: 9241227c45744a901825a20ea8fcc7dfbd768e23