util: refactor create_filtered_rootid
Summary:
# This diff

refactors create_filtered_rootid to use the methods added in the previous diffs

# Context

We have to add support to `eden clone` and `sl clone` for specifying multiple filter paths at clone time. For `eden clone` in particular, we have to provide Eden with some way to generate a V1 FilterId prior to the repository being fully setup. Eden needs this since it sticks an initial FilterId in the checkout's snapshot file. To allow EdenFS to get this information, we'll:

1) Create a generate_filter_id method on FilterGenerator
2) Create a new debug cmd called `debugfilterid` which takes a rev and an arbitrary # of paths. This command will construct a FilterGenerator and return FilterId bytes
3) From `eden clone`, call `sl debugfilterid` on the BackingStore to create a new FilterId
4) Stick the id in the Snapshot file at clone time

Differential Revision: D83801204

fbshipit-source-id: f1bfcd4378fa6da273caf9f393c46a28f37dcee1