clone: support multiple filters
Summary:
# This diff

Refactors `sl clone` to support enabling multiple filters at the same time. This is expected not to work, since `eden clone` hasn't been refactored just yet. However, this demonstrates the fallback logic that will be used in prod while we wait for all users to install new versions of edenfsctl that contain the `--filter-paths` option.

Differential Revision: D83801202

fbshipit-source-id: 09c38a99f6b3c81b6030690891762bec26c59f30