clone: simplify FilteredHg clone logic
Summary:
genevievehelsel pointed out we could reuse the single arg `--filter-path` and add an alias `--filter-paths`. In the future, we can migrate to `--filter-paths`, but for now we'll stick with the old option.

This provides similar behavior to switching to the new flag, but avoids some compatability headaches in the short term.

Reviewed By: genevievehelsel

Differential Revision: D83876469

fbshipit-source-id: b73046b5e567286deb54142f017b65c7ab20dfbc