Define the traversal progress update interval as constant
Summary:
This diff defines the traversal progress update interval as a constant, allowing for easier maintenance and modification in the future. The constant `TRAVERSAL_PROGRESS_UPDATE_INTERVAL` is defined in `types.rs` and is set to 10,000.

Changed the value from 1000 to 10000 because we see traversal speeds now at around 30k files/s and 1000 value was resulting into too frequent progress bar updates.

Reviewed By: akushner

Differential Revision: D83895166

fbshipit-source-id: 268361376193daa797d90da6b344597e854c22fb