Avoid swallowing errors during cache update for tags and content refs
Summary:
When we fail to update the tag cache, we try to log the actual error for the failure. However, for some reason the stack trace actually get's swallowed and we just see the top most error which is the context that we ourselves have added (as can be seen in example below)

```
E1003 02:05:32.397275 1757132 [bonsai_tag_mapping::cache] fbcode/eden/mononoke/repo_attributes/bonsai_tag_mapping/src/cache.rs:111] Failure in receiving notification from tags scribe category. Error: Error while receiving update notification
```

This might be a bug with anyhow so for now I am removing the added context so we might be able to see the actual bug.

Reviewed By: singhsrb

Differential Revision: D83967087

fbshipit-source-id: dd934ac59391267b3f899746684a99beb89c622a