Choose which attribute to show in getAttributesFromFilesV2 debug command
Summary:
# This diff:
Choose which attribute should show up in the getAttributesFromFilesV2 debug command

# Context:
The `getAttributesFromFilesV2` returns the file type which can be `EXECUTABLE` too. However, on Windows it returns `REGULAR` file type for executable files. I want to fix this thrift endpoint to return correct file type on all the platforms which is needed for RE actions. As the first step, I want to create a debug command to test it manually on all the platforms to confirm the future changes.

This is a large command with lots of arguments. The debug cmd class will be large. To make it easier for review, I split the debug command implementations to multiple related changes diffs.

Differential Revision: D83881252

fbshipit-source-id: a489f51e73de793d0add75472473793a244bce77