add more arguments to the debug command
Summary:
# This diff:
Add more arguments to the debug command for scop (files or tree) and sync behavior seconds.

# Context:
The `getAttributesFromFilesV2` returns the file type which can be `EXECUTABLE` too. However, on Windows it returns `REGULAR` file type for executable files. I want to fix this thrift endpoint to return correct file type on all the platforms which is needed for RE actions. As the first step, I want to create a debug command to test it manually on all the platforms to confirm the future changes.

This is a large command with lots of arguments. The debug cmd class will be large. To make it easier for review, I split the debug command implementations to multiple related changes diffs.

Differential Revision: D83885562

fbshipit-source-id: da41d718c1c5a53e2872354475827a741a007774