patch: guess if ParsedDiff represents a submodule change
Summary:
# Context

In the earlier stack, we updated ChangedFile to make it aware of submodule changes, utilizing .gitmodules data present in the working directory.  This is limited to uncommited changes, however, since .gitmodules only reflects the current state.

In ISL, we have other comparison views for committed changes, where the diff data may be directly from `sl diff` and the avobe approach is inapplicable.  While `sl diff` doesn't provide file mode yet, we can still guess based on the pattern of submodule changes.

# This diff
Implemented `guessIsSubmodule` to determine whether a raw diff patch is a submodule change, based on the pattern below:

```
┌─ jldai@devbig279.atn6 /data/users/jldai/aosp-internal (1b31725d)
└─ $ sl diff
 diff --git a/external/brotli b/external/brotli
 --- a/external/brotli
+++ b/external/brotli
@@ -1,1 +1,1 @@
-Subproject commit 892110204ccf44fcd493ae415c9a69c470c2a9cf
+Subproject commit 57de5cc4288565a9c3a7af978ef15f0abf0ada1b
 diff --git a/external/rust/cxx b/external/rust/cxx
 --- a/external/rust/cxx
+++ b/external/rust/cxx
@@ -1,1 +1,1 @@
-Subproject commit 862a23082a087566776280a5b1539d3b62701bcb
+Subproject commit 1869e93e54fa9d9425bd88bdb25073af9ed7e782
```

Reviewed By: evangrayk

Differential Revision: D83791774

fbshipit-source-id: 9defe39147be2eae2159406454aee9e89bac53da