cross_repo_sync: fix implicitly deleting nested submodules
Summary:
While running the intial import in S571506, I consistently hit an issue with
nested submodule implicit deletion as in P1980041282. Turns out the existing
logic only handles implicit deletion for a non nested submodules. I just
extended the same logic as in https://fburl.com/code/lktp5mxi to the nested
submodules and it seems to fix the issue.

Reviewed By: gustavoavena

Differential Revision: D84001035

fbshipit-source-id: 11ba99aa6b1fb680cd470216b58e8f5e2186cbde