ui: add teed buffer mode
Summary:
# This diff

Adds a mode to ui.pushbuffer that allows writes to be both buffered and written to the UI in real time

# Context

Merge driver owners want more information about why their merge drivers are failing. We don't capture UI output for merge drivers, and even if we did, then merge driver output wouldn't be responisve (i.e. we'd buffer it and only print it out when the merge driver finished).

This diff adds a "tee" mode to pushbuffer() that will allow the output from merge drivers to be captured in a buffer, but also simultaneously printed to the UI object so the merge driver output is "responsive"

The captured output can then later be captured in telemetry about merge driver output/failures.

Reviewed By: quark-zju

Differential Revision: D84020507

fbshipit-source-id: 3bc0fe576a8408eb82a688ac5dbce43f675acee0