cgdm: add admin command for updating CGDM
Summary:
Add an admin command for updating `CGDMComponents` blobs. The command finds commits that are ancestors of the provided `heads` and are not already indexed in the `CGDMComponents` blob, and tries to either add them to one of their parent components or create a new component for each. Each commit is allowed to only be added to the parent component with the maximum Id. This enforces a topological order between components sorted by their Ids which we will exploit in the git protocol code.

The CGDM and commit packfile item blobs are only written for components when they become "full" i.e. when either their size reaches the provided max or their byte size reaches the provided max size. This is to avoid writing multiple blobs for every component whenever a new commit gets added and requiring quadratic storage space.

Reviewed By: RajivTS

Differential Revision: D84154782

fbshipit-source-id: 3544e8c201541b9ec8cf7d088f5cee83102313f8