cgdm: add types for CompactedGitDeltaManifest
Summary: CompactedGitDeltaManifest is an optimization of GitDeltaManifest to prevent fetching a large number of small blobs when serving read requests. CGDMComponents is stored in memory and updated periodically, and represents a grouping of commits into components. Every component stores Ids that point to two blobs, one stores the list of all GitDeltaManifests for the component, and the other stores all commit packfiles items.

Reviewed By: RajivTS

Differential Revision: D84149710

fbshipit-source-id: ec4fcc87ce53771decbaa271948239e23a9a948d