Add use-case argument to rust CLI
Summary:
# Context
UseCase is a value used by EdenFsClient to track the origin of calls.
There are many users of EdenFS that interact via the CLI. For these users, they were all getting bucketed under the 'edenfsctl' use case.

# This Diff
Adds a new arg use-case to edenfsctl. This will be used to determine where calls are coming from.

If not passed in, it will use the default EdenFSCtl use case.

Differential Revision: D84183758

fbshipit-source-id: 28c864355a2d13b9094abc6f65c8a34f75b2b909