Show uncaught exceptions in dev mode
Summary:
When in dev mode, it's useful to see top level uncaught exceptions / promise rejections. Otherwise you need to go to the dev tools to find them which takes a moment and is easy to forget to do.

This diff adds code to the vscode webview platform in dev mode to write to the DOM with errors when detected.

Note that this goes around all the normal rendering, so that it can catch top level errors before ANYTHING else loads. This is useful as those are some of the more annoying errors to deal with. This doesn't need to be that pretty since it's just for dev mode.

Reviewed By: nsblake

Differential Revision: D84187818

fbshipit-source-id: 0e618037ce6317abddef32abe824e75bd2e094df