create_repos: Fix bug in the ACL validation logic
Summary:
When we create a repo ACL, we set the provided "ONCALL_GROUP" as maintainer.

If we don't consider `ONCALL_GROUP_TYPE` (a.k.a: "ONCALL_GROUP") as a valid maintainer for the oncall, we don't even recognize the ACLs we create ourselves.

I observed this in prod this morning by:
* Starting a repo creation request including repos that didn't have their ACL set-up yet
* The tool created the ACL with the required "ONCALL_GROUP"
* The mutation failed to land due to a configerator contention issue
* On an attempt to generate the same mutation, I got told the ACL wasn't valid

Reviewed By: mzr

Differential Revision: D84249846

fbshipit-source-id: f60d564d3a47b38af2d4275c5d954183e79fa266