config: use AtomicReadMostlyMainPtr in ReloadableConfig
Summary:
The ReloadableConfig is used in very hot code paths like the ObjectStore after
each aux/blob/tree get, and it thus shows up in profiling. The reason is that
reading a config value is fairly involved and requires reading the timestamp,
comparing that against a point in time, and taking a lock in the fast path. In
the slow path, the config is potentially reloaded from disk.

All of this feels excessive when configs very rarely change. Instead, this diff
takes a different approach: instead of reloading the config when getting a value
from it, let's always reload it in the background and make that reload not block
readers with the help of an AtomicReadMostlyMainPtr. The read path is now
entirely lock free.

Reviewed By: muirdm, MichaelCuevas

Differential Revision: D82133650

fbshipit-source-id: a6159e8345fcfaf304ef17834e144dd249641ef2