Show star next to OD warm bookmark
Summary: When opening an OD, it tags the warmed public commit with sl debusmallcommitmetadata and we render it in ISL. That should be considered "recommended", as it is definitely already warm. So if we have a star next to fbcode/warm, we should also have a star next to OD Warmed Up Commit.

Reviewed By: evanworley, genevievehelsel

Differential Revision: D84176123

fbshipit-source-id: 8934aa3fec0792369740103c590c59cd15ea0386