repository: introduce SubmodulePathCache for server side quick lookup
Summary:
A cache on the server side to enable commands like `openDiff` to quickly determine if a path is a submodule via repo.

(Initially tried to use jotai vanilla store into the `vscode/extension/commands.ts`, and it's proved to be a bad idea.)

`submodulesByRoot` is still the source of truth sync'ed between client and server via subscription.

Differential Revision: D84090147

fbshipit-source-id: 52f6f59001995bf1a2bf2e37abe8ce24a68f1627