Pause getDevEnvType on rebase warnings
Summary:
getDevEnvType is a suspect for flaky issues regarding T239043749. It sends messages to the server and requests IO operations, just for tracking.

Some patterns: multiple affected users were on Devservers (while the function is designed for ODs). From investigating on [Scuba](https://fburl.com/scuba/sapling_isl_analytics/fi849oxi), among the affected people reported seeing/closing the warning, but there are no logs from their unixname for the warnings. This suggests that the runtime made it to the showModal line, but not the the [tracking line](https://fburl.com/code/nkqoooha), which is after the getDevEnvType.

In the future, it may be better to replace this with information about the repository, as these warnings are going to be rolled out to more than just ODs.

Reviewed By: evangrayk

Differential Revision: D84282530

fbshipit-source-id: 6d7bc01ec489e30ddf6ed5e94039c497ce4700e0