Boilerplate for new sidebar panel
Summary: This diff sets up a new webview for showing comments in the sidebar, where we'll render them as a collapsible group instead of a flat list. I figured the easiest way to do this while still maintaining the old webview (i.e. before we roll out the new comments implementation) was to essentially duplicate it.

Reviewed By: evangrayk

Differential Revision: D84190972

fbshipit-source-id: e0c06e8ce8b0088f3b4a9df7b418ef10741acf22