Align name and profile picture
Summary:
As I'm testing the comments sidebar with longer comments, I noticed that the profile picture is being rendered in a separate column from the rest of the comment content. This looks a bit weird since the convention is generally to put the name and profile picture on the same line.

This diff adjusts the flexbox layout so that the comment content better matches how comments are rendered in Phabricator/Diffs. Couple of things to keep in mind with this particular change:

1. This affects both the sidebar and the inset webviews, since they both use this component.
2. This change is currently not behind a GK.

This also more closely aligns with alexchantastic's mockups.

Reviewed By: evangrayk

Differential Revision: D84273463

fbshipit-source-id: ef29c7558e36e656100131eaaae8d6993ee88bca