Support navigating to comment location
Summary:
This diff enables the filepath link on a comment in the sidebar to actually navigate to the location in-file where the comment is rendered, while also expanding the comment.

Fairly straightforward change, but I did need to update the `openFile` method to support links without a selected `Repository` (which it kind of supports anyways?) This enables links from the debug group to work, especially if you're working outside of a repository.

Long-term, although comment groups are designed right now to work without a specified repository, depending on how comments data is received (i.e. filepath format) we may need to consider having some kind of repository reference here.

Reviewed By: evangrayk

Differential Revision: D84273464

fbshipit-source-id: 8c95de45fae4288e59256317f97cf570527bff38